Restructure for APIRest
