+++ src/main/java/apex/ingagers/ecommerce/repository/ProductsRepository.java
@@ -1,9 +1,9 @@
 package apex.ingagers.ecommerce.repository;
 
-import org.springframework.data.repository.CrudRepository;
+import org.springframework.data.jpa.repository.JpaRepository;
 
 import apex.ingagers.ecommerce.model.Products;
 
-public interface ProductsRepository extends CrudRepository<Products, Integer> {
+public interface ProductsRepository extends JpaRepository<Products, Integer> {
     
 }
